requirements.txt

attrs==17.1.0
xmltodict==0.11.0
xmlschema==0.9.7